Restore Linux 32-bit timestamp compilation (#775)
* fix: preserve sync evidence timestamps on linux32

* ci: enforce linux32 sync evidence builds

+++ internal/syncevidence/oneshot_request_times_linux.go
@@ -5,5 +5,5 @@ package syncevidence
 import "golang.org/x/sys/unix"
 
 func oneShotRequestChangeTimes(stat unix.Stat_t) (int64, int64, int64, int64) {
-	return stat.Mtim.Sec, stat.Mtim.Nsec, stat.Ctim.Sec, stat.Ctim.Nsec
+	return int64(stat.Mtim.Sec), int64(stat.Mtim.Nsec), int64(stat.Ctim.Sec), int64(stat.Ctim.Nsec)
 }
